refactor : set the repository implementation on new package implementation.
